Refactor booking system entities and repositories
- Removed Address, Business, and Employee entities as part of restructuring.
- Updated Booking entity to use UUID for bookingId and changed relationships to Organization and User.
- Enhanced BookingStatus enum with descriptive comments.
- Modified Customer entity to use UUID for customerId.
- Updated AddressRepository and BookingRepository to use UUID as the primary key type.
- Removed unused BusinessAddress, BusinessOwner, and Business repositories.
- Refactored BookingService to accommodate changes in entity relationships and repository methods.
- Implemented new OrganizationService with improved error handling and response management.
- Updated application configuration with Keycloak client details.
- Added SQL changes for new booking schema and created new Address entity under Keycloak.
- Added request body JSON for testing organization creation.
- Created a new SQL file for testing queries.

todo:
 - adjust the organization creation process to add current user as a member and assign them as owner

+++ Backend/swift_booking/swift_booking/src/main/java/com/web/app/swift_booking/entity/Keycloak/User.java
@@ -52,6 +52,6 @@ public class User {
     private String serviceAccountClientLink;
 
     @Column(name = "not_before")
-    private Long notBefore;
+    private Integer notBefore;
 
 }
